requirement.txt
langchain==0.0.208
langchain-core<0.3,>=0.1.46
langchain-community==0.0.38
langchain-openai==0.1.6
langchain-anthropic==0.1.13
openai==1.24.0
wikipedia
python-dotenv
pydantic==1.10.12
aiohttp>=3.8.0  # For async support
typing-extensions>=4.2.0  # For enhanced type hints
python-json-logger>=2.0.0  # For structured logging